Not fully working! Append to file working. Now trying to append with form.

+++ images.php
@@ -1,12 +1,6 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: Edwin, Ryan
- * Date: 29/09/2014
- * Time: 11:16 AM
- */
+/* GLOBALS */
 $imageFiles[] = array();
-$textFileName = "images.txt";
 
 /**
  * Returns all files (incl. path) contained within a directory
@@ -34,6 +28,7 @@ function listFolderFiles( $currDir, &$imageFiles )
     return $imageFiles;
 }
 
+
 /**
  * @param $imageFiles
  * @return array filtered list of images
@@ -48,28 +43,74 @@ function FilterFileList( $imageFiles )
     return $targetFiles;
 }
 
+
+function appendToFile($path, $name, $text_Input)
+{
+    $file = 'images.txt';
+    // Open the file to get existing content
+    $current = file_get_contents($file);
+    // Append a new record to the file
+    $current .= "$path\t$name\t$text_Input\n";
+    // Write the contents back to the file
+    file_put_contents($file, $current);
+}
+
+function isInFile ($theUnknown)
+{
+    $file = 'images.txt';
+    if( strpos(file_get_contents($file),$theUnknown) !== false)
+        return true; // somewhere in the file
+    else return false;
+}
+?>
+
+
+<?php
 $filteredFileList = FilterFileList( listFolderFiles( $dir, $imageFiles ) );
-$_SESSION['TaggedImages'] = $filteredFileList;
-WriteTaggedImagesToFile($textFileName);
 
-/* Show image and ask for input */
-foreach ($filteredFileList as $photo) {
-    echo ' <img src=" ' . $photo . ' " width="400"/><br />';
+// Create array of UntaggedImages from directory and cross check
+$theUntagged = array();
+// For each item not in the file, add to theUntagged
+foreach ($filteredFileList as $theUnknown) {
+    if ( !isInFile(dirname($theUnknown) . "\t" . basename($theUnknown)) ) {
+        $theUntagged[] = $theUnknown; // Add to theUntagged
+    }
 }
 
-function WriteTaggedImagesToFile($textFileName) {
-    $fp = fopen($textFileName, 'a+') or die("Cannot open file.");
-    fwrite($textFileName, $_SESSION['TaggedImages']);
-    fclose($fp);
-}
+if ( !empty($theUntagged) ) {
+    echo ' <img src=" ' . $theUntagged[0] . ' " width="400"/><br />';
+    $theUntaggedString = (string)$theUntagged[0];
+    $path = dirname($theUntaggedString);
+    $name = basename($theUntaggedString);
+    ?>
 
-/* TODO: come back later. This is to read from file.
- * ReadTextFile($textFileName);
+    <form class="" id="" action="index.php" method="post">
+        <input type="text" name="path" value="/">
+        <input type="hidden" name="name" value="<?php echo "$name"; ?>">
+        <input type="text" id="humanInput" name="humanInput" aria-label="people" placeholder="people" width="700px">
+        <input type="submit" value="Save">
+    </form>
 
-function ReadTextFile( $textFileName ) {
-    $fp = fopen($textFileName, 'a+') or die("Cannot open file.");
-    //Create array to hold temporary tagged images and their path
-    $tempImagePaths = array();
-    $tempImageFiles = array();
 
-}*/
+<!--/* Show image and ask for input */
+foreach ($_SESSION['UntaggedImages'] as $photo) {
+    // Check if image is already in text file
+    $alreadyThere = false; // TODO: Check here
+    $humanInput = "Placeholder"; // TODO: Input field here
+    // if there already skip
+    if ($alreadyThere) {
+        // skip to next image
+        echo ' <img src=" ' . $photo . ' " width="400"/><br />';
+        // else wait for human to click next and call appendToFile
+    } else
+    {
+        echo ' <img src=" ' . $photo . ' " width="400"/><br />';
+        ?>
+
+    <form class="" id="" action="index.php" method="post">
+        <input type="text" id="humanInput" name="humanInput" aria-label="people" placeholder="people" width="700px">
+        <input type="submit" value="Save">
+    </form>-->
+<?php
+}   //appendToFile(dirname($photo), basename($photo), $humanInput);
+?>
